More fixes for Chardonnay testnet (#273)
* move AxelarBody back to []byte

* axelarcork validator cap, check cork count

* fix axelarcork module account sweep

* Make sure cellar IDs are always sorted

* sort cellar IDs when written to state

* Always store fee accrual counters sorted

* always sort allowed addresses for publisher intents

* update v7 upgrade with new cellars

* appease linter

+++ x/axelarcork/types/axelar.go
@@ -10,7 +10,7 @@ const (
 type AxelarBody struct {
 	DestinationChain   string `json:"destination_chain"`
 	DestinationAddress string `json:"destination_address"`
-	Payload            []int  `json:"payload"`
+	Payload            []byte `json:"payload"`
 	Type               int64  `json:"type"`
 	Fee                *Fee   `json:"fee"`
 }
